pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpfuzz"
version = "0.1.0"
description = "Dynamic security testing for MCP servers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Ven" }]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mcp>=1.0",
]

[project.scripts]
mcpfuzz = "mcpfuzz.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
